refactor(web): consolidate date formatters, rename utils with prefix convention
- Extract formatFriendlyDate to utils/format-date.ts, replacing the
  domain-specific formatLibraryDate and consolidating with the
  invoices-modal formatDate (both were near-identical toLocaleDateString
  wrappers).
- Rename feed-utils.ts → utils-feed.ts to follow the utils- prefix
  convention for domain utility files (groups them in the file tree).
- Remove domains/library/library-date.ts (dead after consolidation).
- Add CONVENTIONS.md sections: 'No single-file directories' and
  'Utility file naming within domains' documenting the patterns.
