--- apps/web/src/domains/library/library-date.ts
@@ -1,8 +1,0 @@
-export function formatLibraryDate(epochMs: number): string {
-  const date = new Date(epochMs);
-  return date.toLocaleDateString(undefined, {
-    day: "numeric",
-    month: "short",
-    year: date.getFullYear() !== new Date().getFullYear() ? "numeric" : undefined,
-  });
-}
